requirements.txt
feedparser>=6.0.10
requests>=2.31.0
openai>=1.0.0
trafilatura>=1.6.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
